+++ modules/entry/src/main/java/com/thekeeperofpie/artistalleydatabase/entry/grid/EntryGrid.kt
@@ -309,6 +309,8 @@ object EntryGrid {
                         contentDescription = stringResource(R.string.selected_content_description),
                     )
                 }
+
+                entry.ErrorIcons(modifier = Modifier.align(Alignment.BottomEnd))
             }
         }
     }