+++ modules/entry/src/main/java/com/thekeeperofpie/artistalleydatabase/entry/grid/EntryGridModel.kt
@@ -1,6 +1,8 @@
 package com.thekeeperofpie.artistalleydatabase.entry.grid
 
 import android.net.Uri
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.thekeeperofpie.artistalleydatabase.entry.EntryId
 
 interface EntryGridModel {
@@ -10,4 +12,11 @@ interface EntryGridModel {
     val imageWidth: Int?
     val imageHeight: Int?
     val imageWidthToHeightRatio: Float
+
+    /**
+     * Indicator icons to show to the user on top of the grid cell. "Error" isn't quite the right
+     * term to use here, but "indicator" isn't, either.
+     */
+    @Composable
+    fun ErrorIcons(modifier: Modifier)
 }